pyproject.toml
[project]
name = "gh-wx-tool"
version = "1.0.0"
description = "WeChat database decryption and message search tool"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "cryptography>=41.0.0",
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "zstandard>=0.23.0",
    "pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/wechat_decrypt_tool"]

[tool.uv]
dev-dependencies = []

[tool.ruff]
line-length = 120
target-version = "py38"
